Made env. var file backup dependent upon whether it existed or not

+++ bash/stack/update_environment_vars.sh
@@ -6,12 +6,16 @@ source $PRODIGUER_HOME/bash/init.sh
 # Main entry point.
 main()
 {
-	declare backup=$HOME/.prodiguer-"`date +%Y-%m-%d`"
-	cp $HOME/.prodiguer $backup
+	log "UPDATING ENVIRONMENT VARS ..."
+
+	if [ -f $HOME/.prodiguer ]; then
+		declare backup=$HOME/.prodiguer-"`date +%Y-%m-%d`"
+		cp $HOME/.prodiguer $backup
+		log "previous environment variable configuration file backup @ "$backup
+	fi
 	cp $PRODIGUER_DIR_TEMPLATES/prodiguer_env.sh $HOME/.prodiguer
 
 	log "UPDATED ENVIRONMENT VARS"
-	log "previous configuration file backup @ "$backup
 }
 
 # Invoke entry point.
